Moves util functions to separate util files.
