--- src/utils.cpp
@@ -1,111 +1,0 @@
-#include "utils.hpp"
-#include "exceptions.hpp"
-#include <opencv2/core.hpp>
-#include <opencv2/core/eigen.hpp>
-#include <opencv2/imgcodecs.hpp>
-#include <unordered_set>
-
-using namespace utils;
-
-#pragma region Matrices
-Eigen::MatrixXd matrix::fromJson(const json &values) {
-  if (values.empty()) {
-    return {};
-  }
-
-  if (!values.is_array()) {
-    throw exceptions::json::JSONTypeException();
-  }
-  if (!values[0].is_array()) {
-    throw exceptions::json::JSONArray2DException();
-  }
-
-  int rows = values.size(), cols = values[0].size();
-  Eigen::MatrixXd result(rows, cols);
-
-  for (int i = 0; i < rows; ++i) {
-    for (int j = 0; j < cols; ++j) {
-      if (values[i][j].is_array()) {
-        throw exceptions::json::JSONArray2DException();
-      }
-      if (!values[i][j].is_number()) {
-        throw exceptions::json::JSONTypeException();
-      }
-      result(i, j) = values[i][j];
-    }
-  }
-
-  return result;
-}
-
-Eigen::MatrixXd utils::matrix::flatten(const Eigen::MatrixXd &in) {
-  Eigen::MatrixXd out(in);
-  return out.reshaped<Eigen::RowMajor>().transpose();
-}
-#pragma endregion Matrices
-
-#pragma region Math
-Eigen::MatrixXi utils::math::oneHotEncode(const std::vector<int> &targets,
-                                          int numClasses) {
-  Eigen::MatrixXi result = Eigen::MatrixXi::Zero(targets.size(), numClasses);
-  for (int i = 0; i < targets.size(); ++i) {
-    if (targets[i] >= numClasses || targets[i] < 0) {
-      throw exceptions::utils::one_hot_encode::InvalidLabelIndexException();
-    }
-    result(i, targets[i]) = 1;
-  }
-  return result;
-}
-
-Eigen::MatrixXd utils::math::normalise(const Eigen::MatrixXd &data,
-                                       std::pair<float, float> from,
-                                       std::pair<float, float> to) {
-  auto [fromMin, fromMax] = from;
-  if (fromMin >= fromMax) {
-    throw exceptions::utils::normalise::InvalidRangeException();
-  }
-
-  auto [toMin, toMax] = to;
-  if (toMin >= toMax) {
-    throw exceptions::utils::normalise::InvalidRangeException();
-  }
-
-  return (data.array() - fromMin) * (toMax - toMin) / (fromMax - fromMin) +
-         toMin;
-}
-#pragma endregion Math
-
-#pragma region Path
-std::vector<std::filesystem::path>
-utils::path::glob(const std::filesystem::path &path,
-                  std::vector<std::string> extensions) {
-  std::unordered_set<std::string> extensionSet(extensions.begin(),
-                                               extensions.end());
-  std::vector<std::filesystem::path> result;
-  for (auto const &entry :
-       std::filesystem::recursive_directory_iterator(path)) {
-    if (!std::filesystem::is_directory(entry.path()) &&
-        extensionSet.count(entry.path().extension())) {
-      result.push_back(entry.path());
-    }
-  };
-  return result;
-};
-#pragma endregion Path
-
-#pragma region Image
-Eigen::MatrixXd utils::image::openAsMatrix(std::filesystem::path path) {
-  Eigen::MatrixXd result;
-  cv::Mat img = cv::imread(path, cv::IMREAD_UNCHANGED);
-  if (img.empty()) {
-    throw exceptions::utils::image::InvalidImageFileException(path);
-  }
-  cv::cv2eigen(img, result);
-  return result;
-}
-
-Eigen::MatrixXd utils::image::normalise(const Eigen::MatrixXd &data) {
-  return utils::math::normalise(data, std::make_pair(0, 255),
-                                std::make_pair(-1, 1));
-}
-#pragma endregion Image
