+++ src/utils/all.hpp
@@ -1,0 +1,4 @@
+#include "image.hpp"
+#include "math.hpp"
+#include "matrix.hpp"
+#include "path.hpp"
