+++ src/utils/image.cpp
@@ -1,0 +1,21 @@
+#include "image.hpp"
+#include "../exceptions.hpp"
+#include "math.hpp"
+#include <opencv2/core.hpp>
+#include <opencv2/core/eigen.hpp>
+#include <opencv2/imgcodecs.hpp>
+
+Eigen::MatrixXd utils::image::openAsMatrix(std::filesystem::path path) {
+  Eigen::MatrixXd result;
+  cv::Mat img = cv::imread(path, cv::IMREAD_UNCHANGED);
+  if (img.empty()) {
+    throw exceptions::utils::image::InvalidImageFileException(path);
+  }
+  cv::cv2eigen(img, result);
+  return result;
+}
+
+Eigen::MatrixXd utils::image::normalise(const Eigen::MatrixXd &data) {
+  return utils::math::normalise(data, std::make_pair(0, 255),
+                                std::make_pair(-1, 1));
+}
